Adds Partial Output To ValidationContext
Adds the value that will be returned by `validate` to the
ValidationContext, so inline validators can base their output and
validity on previously processed properties.

Adds `is.asyncFunc` and `is.promise` to the TypeScript definitions.

Updates documentation for the ValidationContext, and to all TypeScript
interfaces.

Reduces verbosity when returning IValueOrError, so consumers don't have
to set a bunch of null properties. Blueprint guarantees them for us, so
no reason to set them explicitly.